Created util_data_loader.py
